DBpedia Method
				                                           README DBpedia
It contains three major modules:
1)	Entity Identification and linking and grouping:
Step1: It demands DBpedia Spotlight installation in the system to open the server.
The command : java -Xms512m -Xmx6000m -d64 -jar dbpedia-spotlight-latest.jar en http://localhost:2222/rest

Step2:  Run Split.java: This program splits the paragraph corpus in text into many smaller size
text file.

It demands two paths: 1) The path of input file (paragraph corpus in text)
		          2) The output file location where the split files will be stored.
Grouping:
For example:     Country    <USA, India, China, Mexico>

Step3: It demands the installation of curl command in the command line.
Step4 : Run ReadandProcessFile.java : This class is used to read data chunks of paragraph corpus identifies the keyword and its category.
It demands two paths: 1) The path of input file (the output path of step2)
		          2) The output file location where the entity link data can be stored.


2)	  Page Ranking: Each keyword and category refers to its DBpedia page. The keyword page gives     
reference of the category page. The concept is similar to a page giving citation of the other page.
The score is proportional to the number of reference.
---------The above process till step4 is the indexing process to get the entity keyword and category relationship  to generate a runfile--------------------------------------------------
              
	     The runfile name : curlout.txt      
	      Step5: Run Index.java: The grouping module collect all the related entity under it. This program    
            identifies to which page has got maximum reference and citation and accordingly score them and       
            rank them.
It demands two paths: 1) The path of input file (the output path of step4)
		          2) The output file location where the ranking can be stored.

Constraints: Since it demands DBpedia spotlight installation and curl command installation. I have provided the input file. I will provide the input file of step5.
The file name is curloutput.txt (It is the runfie generated after all the linking has been done it is categorized.)